drop_sources_c.rs hands its krec source generator to drop_sources_d.rs, a pure text move under the 800-line ceiling
