+++ crates/almide-mir/src/lower/drop_sources_c.rs
@@ -553,655 +553,3 @@ fn emit_demanded_shared_drops(
     // is still computed above (by `record_drop_field_frees`) purely to preserve that
     // function's shared signature with the anon-record caller; its value is unused here.
 }
-
-/// The C-015 STRING-FIELD-record key/element twins (`__krec_*`) — generated per
-/// record shape used as a Map key / Set element / `list.unique` element anywhere
-/// in the program. The key normalizes INJECTIVELY into a String (each String
-/// field length-prefixed `<len>:<bytes>,`, each scalar field `<digits>,` — the
-/// netstring discipline, so distinct field values can never collide), and the
-/// backing container is the proven `_str`/`_skv` family; `krec_call_name`
-/// (control_p2.rs) routes the call sites to these names. Over-generation is
-/// harmless (a shape whose call never fires leaves inert fns); a record with a
-/// non-String/scalar field is never collected (its calls keep their wall).
-pub fn generate_krec_sources(
-    program: &almide_ir::IrProgram,
-    type_decls: &[almide_ir::IrTypeDecl],
-) -> String {
-    use almide_ir::visit::{walk_expr, IrVisitor};
-    use almide_lang::types::constructor::TypeConstructorId;
-
-    // Admissible record decls: name -> declaration-ordered field types.
-    let recs: std::collections::HashMap<String, Vec<Ty>> = type_decls
-        .iter()
-        .filter_map(|d| match &d.kind {
-            almide_ir::IrTypeDeclKind::Record { fields } => {
-                let tys: Vec<Ty> = fields.iter().map(|f| f.ty.clone()).collect();
-                (!tys.is_empty()
-                    && tys.iter().all(|t| matches!(t, Ty::Int | Ty::Bool | Ty::String))
-                    && tys.iter().any(|t| matches!(t, Ty::String)))
-                .then(|| (d.name.as_str().to_string(), tys))
-            }
-            _ => None,
-        })
-        .collect();
-    if recs.is_empty() {
-        return String::new();
-    }
-
-    #[derive(Default)]
-    struct Uses {
-        map_iv: std::collections::BTreeSet<String>,
-        map_sv: std::collections::BTreeSet<String>,
-        sets: std::collections::BTreeSet<String>,
-        uniques: std::collections::BTreeSet<String>,
-        /// STRUCTURAL record element shapes (anon hash -> field types, SOURCE order).
-        uniq_structs: std::collections::BTreeMap<String, Vec<Ty>>,
-    }
-    struct Scan<'a> {
-        recs: &'a std::collections::HashMap<String, Vec<Ty>>,
-        uses: Uses,
-    }
-    impl Scan<'_> {
-        /// Pattern-1/2 split (codopsy8 complexity sweep): the 3 groups below are
-        /// independent, self-contained classifications of `ty` (a Map/Set/List shape are
-        /// mutually exclusive, so calling all 3 in sequence is behaviorally identical to
-        /// the original `match ty { .. }`, which already had a `_ => {}` fallback — no
-        /// exhaustiveness guarantee lost). Pure text-move, no logic change.
-        fn note(&mut self, ty: &Ty) {
-            self.note_map(ty);
-            self.note_set(ty);
-            self.note_list(ty);
-        }
-
-        /// Extracted from `note` (codopsy8 complexity sweep, group 1 of 3): `Map[<record>,
-        /// Int/Bool/String]` — the `_iv`/`_sv` value-class split. Verbatim.
-        fn note_map(&mut self, ty: &Ty) {
-            let Ty::Applied(TypeConstructorId::Map, a) = ty else { return };
-            if a.len() != 2 {
-                return;
-            }
-            let Ty::Named(n, _) = &a[0] else { return };
-            if !self.recs.contains_key(n.as_str()) {
-                return;
-            }
-            match &a[1] {
-                Ty::Int | Ty::Bool => {
-                    self.uses.map_iv.insert(n.as_str().to_string());
-                }
-                Ty::String => {
-                    self.uses.map_sv.insert(n.as_str().to_string());
-                }
-                _ => {}
-            }
-        }
-
-        /// Extracted from `note` (codopsy8 complexity sweep, group 2 of 3): `Set[<record>]`.
-        /// Verbatim.
-        fn note_set(&mut self, ty: &Ty) {
-            let Ty::Applied(TypeConstructorId::Set, a) = ty else { return };
-            if a.len() != 1 {
-                return;
-            }
-            if let Ty::Named(n, _) = &a[0] {
-                if self.recs.contains_key(n.as_str()) {
-                    self.uses.sets.insert(n.as_str().to_string());
-                }
-            }
-        }
-
-        /// Extracted from `note` (codopsy8 complexity sweep, group 3 of 3): `List[<record>]`
-        /// (`list.unique` element) — both the DECLARED-record case and an UNANNOTATED
-        /// literal's STRUCTURAL record element (keyed by the anon hash, fields in the
-        /// block's SOURCE order — the r5 lesson). Verbatim.
-        fn note_list(&mut self, ty: &Ty) {
-            let Ty::Applied(TypeConstructorId::List, a) = ty else { return };
-            if a.len() != 1 {
-                return;
-            }
-            if let Ty::Named(n, _) = &a[0] {
-                if self.recs.contains_key(n.as_str()) {
-                    self.uses.uniques.insert(n.as_str().to_string());
-                }
-            }
-            if let Ty::Record { fields } = &a[0] {
-                if !fields.is_empty()
-                    && fields.iter().all(|(_, t)| matches!(t, Ty::Int | Ty::Bool | Ty::String))
-                    && fields.iter().any(|(_, t)| matches!(t, Ty::String))
-                {
-                    self.uses.uniq_structs.insert(
-                        crate::lower::anon_record_drop_name(fields),
-                        fields.iter().map(|(_, t)| t.clone()).collect(),
-                    );
-                }
-            }
-        }
-    }
-    impl IrVisitor for Scan<'_> {
-        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-            self.note(&e.ty);
-            walk_expr(self, e);
-        }
-    }
-    let mut scan = Scan { recs: &recs, uses: Uses::default() };
-    for f in program
-        .functions
-        .iter()
-        .chain(program.modules.iter().flat_map(|m| m.functions.iter()))
-    {
-        for p in &f.params {
-            scan.note(&p.ty);
-        }
-        scan.note(&f.ret_ty);
-        almide_ir::visit::IrVisitor::visit_expr(&mut scan, &f.body);
-    }
-    let uses = scan.uses;
-    if uses.map_iv.is_empty()
-        && uses.map_sv.is_empty()
-        && uses.sets.is_empty()
-        && uses.uniques.is_empty()
-        && uses.uniq_structs.is_empty()
-    {
-        return String::new();
-    }
-
-    let mut out = String::new();
-    let mut norm_emitted: std::collections::BTreeSet<String> = Default::default();
-    let mut emit_norm_tys = |out: &mut String, r: &str, tys: &[Ty]| {
-        if !norm_emitted.insert(r.to_string()) {
-            return;
-        }
-        out.push_str(&format!("fn __krec_norm_{r}(k: Value) -> String = {{\n"));
-        out.push_str("  let h = prim.handle(k)\n");
-        out.push_str("  let a0 = \"\"\n");
-        for (i, t) in tys.iter().enumerate() {
-            let off = 12 + 8 * i;
-            let prev = format!("a{i}");
-            let cur = format!("a{}", i + 1);
-            if matches!(t, Ty::String) {
-                out.push_str(&format!(
-                    "  let s{i}: String = prim.load_str(h + {off})\n  \
-                     let {cur} = {prev} + int.to_string(string.len(s{i})) + \":\" + s{i} + \",\"\n"
-                ));
-            } else {
-                out.push_str(&format!(
-                    "  let {cur} = {prev} + int.to_string(prim.load64(h + {off})) + \",\"\n"
-                ));
-            }
-        }
-        out.push_str(&format!("  a{}\n}}\n", tys.len()));
-    };
-
-    // `r` is a record NAME (`recs`' HashMap key) — a cross-module record carries
-    // its dotted module prefix (`m.Cfg`), which is only valid Almide syntax as a
-    // TYPE reference. Every `__krec_*` string below uses it as a FUNCTION NAME, so
-    // each loop derives the sanitized `rf` (via `drop_fn_ident`, the same dots→
-    // underscores mangling `generate_record_drop_sources` applies) and formats
-    // with `{rf}`, keeping `r`/`recs[r]` only for the HashMap lookup.
-    for r in uses.map_iv.iter() {
-        let rf = drop_fn_ident(r);
-        emit_norm_tys(&mut out, &rf, &recs[r]);
-        out.push_str(&format!(
-            "fn __krec_mfl_{rf}_iv_at(pairs: List[(Value, Int)], i: Int, m: Map[String, Int]) -> Map[String, Int] =\n  \
-               if i >= list.len(pairs) then m\n  \
-               else match list.get(pairs, i) {{\n    \
-                 some(p) => {{\n      \
-                   let (k, v) = p\n      \
-                   __krec_mfl_{rf}_iv_at(pairs, i + 1, map.set(m, __krec_norm_{rf}(k), v))\n    }},\n    \
-                 none => m,\n  }}\n\
-             fn __krec_map_from_list_{rf}_iv(pairs: List[(Value, Int)]) -> Map[String, Int] = {{\n  \
-               let m: Map[String, Int] = map.new()\n  \
-               __krec_mfl_{rf}_iv_at(pairs, 0, m)\n}}\n\
-             fn __krec_map_set_{rf}_iv(m: Map[String, Int], k: Value, v: Int) -> Map[String, Int] =\n  \
-               map.set(m, __krec_norm_{rf}(k), v)\n\
-             fn __krec_map_get_{rf}_iv(m: Map[String, Int], k: Value) -> Option[Int] =\n  \
-               map.get(m, __krec_norm_{rf}(k))\n\
-             fn __krec_map_contains_{rf}_iv(m: Map[String, Int], k: Value) -> Bool =\n  \
-               map.contains(m, __krec_norm_{rf}(k))\n"
-        ));
-    }
-    for r in uses.map_sv.iter() {
-        let rf = drop_fn_ident(r);
-        emit_norm_tys(&mut out, &rf, &recs[r]);
-        out.push_str(&format!(
-            "fn __krec_mfl_{rf}_sv_at(pairs: List[(Value, String)], i: Int, m: Map[String, String]) -> Map[String, String] =\n  \
-               if i >= list.len(pairs) then m\n  \
-               else match list.get(pairs, i) {{\n    \
-                 some(p) => {{\n      \
-                   let (k, v) = p\n      \
-                   __krec_mfl_{rf}_sv_at(pairs, i + 1, map.set(m, __krec_norm_{rf}(k), v))\n    }},\n    \
-                 none => m,\n  }}\n\
-             fn __krec_map_from_list_{rf}_sv(pairs: List[(Value, String)]) -> Map[String, String] = {{\n  \
-               let m: Map[String, String] = map.new()\n  \
-               __krec_mfl_{rf}_sv_at(pairs, 0, m)\n}}\n\
-             fn __krec_map_set_{rf}_sv(m: Map[String, String], k: Value, v: String) -> Map[String, String] =\n  \
-               map.set(m, __krec_norm_{rf}(k), v)\n\
-             fn __krec_map_get_{rf}_sv(m: Map[String, String], k: Value) -> Option[String] =\n  \
-               map.get(m, __krec_norm_{rf}(k))\n\
-             fn __krec_map_contains_{rf}_sv(m: Map[String, String], k: Value) -> Bool =\n  \
-               map.contains(m, __krec_norm_{rf}(k))\n"
-        ));
-    }
-    for r in uses.sets.iter() {
-        let rf = drop_fn_ident(r);
-        emit_norm_tys(&mut out, &rf, &recs[r]);
-        out.push_str(&format!(
-            "fn __krec_sfl_{rf}_at(xs: List[Value], i: Int, acc: Set[String]) -> Set[String] =\n  \
-               if i >= list.len(xs) then acc\n  \
-               else match list.get(xs, i) {{\n    \
-                 some(x) => __krec_sfl_{rf}_at(xs, i + 1, set.insert(acc, __krec_norm_{rf}(x))),\n    \
-                 none => acc,\n  }}\n\
-             fn __krec_set_from_list_{rf}(xs: List[Value]) -> Set[String] = {{\n  \
-               let acc: Set[String] = set.new()\n  \
-               __krec_sfl_{rf}_at(xs, 0, acc)\n}}\n\
-             fn __krec_set_insert_{rf}(s: Set[String], x: Value) -> Set[String] = set.insert(s, __krec_norm_{rf}(x))\n\
-             fn __krec_set_contains_{rf}(s: Set[String], x: Value) -> Bool = set.contains(s, __krec_norm_{rf}(x))\n"
-        ));
-    }
-    for (hash, tys) in uses.uniq_structs.iter() {
-        emit_norm_tys(&mut out, hash, tys);
-        let r = hash;
-        out.push_str(&format!(
-            "fn __krec_uniqfill_{r}(h: Int, oh: Int, n: Int, i: Int, cnt: Int, seen: Set[String]) -> Int =\n  \
-               if i >= n then cnt\n  \
-               else {{\n    \
-                 let x: Value = prim.load_handle(h + 12 + i * 8)\n    \
-                 let key = __krec_norm_{r}(x)\n    \
-                 if set.contains(seen, key) then __krec_uniqfill_{r}(h, oh, n, i + 1, cnt, seen)\n    \
-                 else {{\n      \
-                   let e = prim.load64(h + 12 + i * 8)\n      \
-                   prim.rc_inc(e)\n      \
-                   prim.store64(oh + 12 + cnt * 8, e)\n      \
-                   __krec_uniqfill_{r}(h, oh, n, i + 1, cnt + 1, set.insert(seen, key))\n    }}\n  }}\n\
-             fn __krec_list_unique_{r}(xs: List[Value]) -> List[Value] = {{\n  \
-               let h = prim.handle(xs)\n  \
-               let n = prim.load32(h + 4)\n  \
-               let out: List[Value] = prim.alloc_list_str(n)\n  \
-               let seen: Set[String] = set.new()\n  \
-               let cnt = __krec_uniqfill_{r}(h, prim.handle(out), n, 0, 0, seen)\n  \
-               prim.store32(prim.handle(out) + 4, cnt)\n  \
-               out\n}}\n"
-        ));
-    }
-    for r in uses.uniques.iter() {
-        let rf = drop_fn_ident(r);
-        emit_norm_tys(&mut out, &rf, &recs[r]);
-        out.push_str(&format!(
-            "fn __krec_uniqfill_{rf}(h: Int, oh: Int, n: Int, i: Int, cnt: Int, seen: Set[String]) -> Int =\n  \
-               if i >= n then cnt\n  \
-               else {{\n    \
-                 let x: Value = prim.load_handle(h + 12 + i * 8)\n    \
-                 let key = __krec_norm_{rf}(x)\n    \
-                 if set.contains(seen, key) then __krec_uniqfill_{rf}(h, oh, n, i + 1, cnt, seen)\n    \
-                 else {{\n      \
-                   let e = prim.load64(h + 12 + i * 8)\n      \
-                   prim.rc_inc(e)\n      \
-                   prim.store64(oh + 12 + cnt * 8, e)\n      \
-                   __krec_uniqfill_{rf}(h, oh, n, i + 1, cnt + 1, set.insert(seen, key))\n    }}\n  }}\n\
-             fn __krec_list_unique_{rf}(xs: List[Value]) -> List[Value] = {{\n  \
-               let h = prim.handle(xs)\n  \
-               let n = prim.load32(h + 4)\n  \
-               let out: List[Value] = prim.alloc_list_str(n)\n  \
-               let seen: Set[String] = set.new()\n  \
-               let cnt = __krec_uniqfill_{rf}(h, prim.handle(out), n, 0, 0, seen)\n  \
-               prim.store32(prim.handle(out) + 4, cnt)\n  \
-               out\n}}\n"
-        ));
-    }
-    out
-}
-
-/// The `Result[(V1, V2), String]` VARIANT-PAIR wrapper drops (#1547 shape 1 —
-/// the aggregate-transition return `(new_state, event)`): for every such
-/// Result type the program USES, generate `$__drop_vp_<A>_<B>` — the pair
-/// routine `DropWrapperRec` recurses into for the Ok payload (Err is the flat
-/// @12 String the wrapper render already decs). Per slot the free follows the
-/// STRUCTURAL rule the field-frees generators use: a RICH variant recurses via
-/// its generated `$__drop_<V>`; a FLAT variant / String / Bytes is
-/// one-level-exact — one `rc_dec` is its full free. Usage-driven like
-/// `generate_krec_sources` (a per-pair emission over every declared pair
-/// would be quadratic); trusted prim-only routines, leak-loop class.
-pub fn generate_variant_pair_result_sources(
-    program: &almide_ir::IrProgram,
-    type_decls: &[almide_ir::IrTypeDecl],
-) -> String {
-    use almide_ir::visit::walk_expr;
-    use almide_lang::types::constructor::TypeConstructorId;
-
-    let variant_names = variant_type_names(type_decls);
-    let flat_names = flat_variant_type_names(type_decls);
-    let all_record_names: std::collections::HashSet<String> = type_decls
-        .iter()
-        .filter(|d| matches!(&d.kind, almide_ir::IrTypeDeclKind::Record { .. }))
-        .map(|d| d.name.as_str().to_string())
-        .collect();
-    let rich: std::collections::HashSet<String> = type_decls
-        .iter()
-        .filter(|d| variant_needs_recursive_drop(d, &variant_names, &all_record_names))
-        .map(|d| d.name.as_str().to_string())
-        .collect();
-    // RECORD slots complete the #1564 matrix (#1547 shape 1 was closed on the
-    // variant cell only — an ordinary aggregate is a record): a RECURSIVE-drop
-    // record recurses via its unconditionally-generated `$__drop_<R>` (the
-    // same `__drop_<ident>` spelling the rich-variant arm emits), an
-    // ALL-SCALAR record is one-level-exact under a flat rc_dec.
-    let rich_recs: std::collections::HashSet<String> =
-        recursive_record_drop_names(type_decls).into_iter().collect();
-    let flat_recs: std::collections::HashSet<String> = type_decls
-        .iter()
-        .filter_map(|d| match &d.kind {
-            almide_ir::IrTypeDeclKind::Record { fields }
-                if fields.iter().all(|f| !crate::lower::is_heap_ty(&f.ty)) =>
-            {
-                Some(d.name.as_str().to_string())
-            }
-            _ => None,
-        })
-        .collect();
-
-    // Is `t` an admissible pair SLOT, and how does it free? `Some(true)` =
-    // recurse via the generated `$__drop_<V>` / `$__drop_<R>`, `Some(false)` =
-    // flat rc_dec.
-    let slot_class = |t: &Ty| -> Option<bool> {
-        match t {
-            Ty::Named(n, args) if args.is_empty() => {
-                let n = n.as_str();
-                if rich.contains(n) || rich_recs.contains(n) {
-                    Some(true)
-                } else if flat_names.contains(n) || flat_recs.contains(n) {
-                    Some(false)
-                } else {
-                    None
-                }
-            }
-            Ty::String | Ty::Bytes => Some(false),
-            // LIST slots (#1580): a scalar-element list frees flat (one
-            // rc_dec); a `List[String]` frees per-element via the vp-private
-            // `__drop_vp_list_str` (rich). Deeper element classes decline —
-            // the ctor gate mirrors this exactly.
-            Ty::Applied(TypeConstructorId::List, la) if la.len() == 1 => {
-                if !crate::lower::is_heap_ty(&la[0]) {
-                    Some(false)
-                } else if matches!(la[0], Ty::String) {
-                    Some(true)
-                } else {
-                    None
-                }
-            }
-            // A SCALAR slot (#1579's mixed pair, `(Int, Note("a", n))`): the
-            // slot holds a raw value, not a handle — the emitted drop SKIPS
-            // it entirely (see `slot_free`; an rc_dec would dec a non-handle).
-            _ if !crate::lower::is_heap_ty(t) => Some(false),
-            _ => None,
-        }
-    };
-
-    // Collect every used pair, deterministically ordered.
-    struct Finder<'a> {
-        pairs: std::collections::BTreeSet<(String, String)>,
-        slot_class: &'a dyn Fn(&Ty) -> Option<bool>,
-    }
-    impl Finder<'_> {
-        fn check(&mut self, ty: &Ty) {
-            if let Ty::Applied(TypeConstructorId::Result, a) = ty {
-                if a.len() == 2 && matches!(a[1], Ty::String) {
-                    if let Ty::Tuple(ts) = &a[0] {
-                        if ts.len() == 2
-                            && (self.slot_class)(&ts[0]).is_some()
-                            && (self.slot_class)(&ts[1]).is_some()
-                            // At least one RICH slot — an all-flat pair frees
-                            // exactly like (String, Int)'s existing routes.
-                            && ((self.slot_class)(&ts[0]) == Some(true)
-                                || (self.slot_class)(&ts[1]) == Some(true))
-                        {
-                            self.pairs.insert((ty_slot_name(&ts[0]), ty_slot_name(&ts[1])));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    fn ty_slot_name(t: &Ty) -> String {
-        use almide_lang::types::constructor::TypeConstructorId;
-        match t {
-            Ty::Named(n, _) => n.as_str().to_string(),
-            Ty::String => "String".to_string(),
-            Ty::Bytes => "Bytes".to_string(),
-            // The lowercase spellings are the ctor gate's reserved slot names
-            // (`variant_pair_result_drop_fn`): type names are
-            // Uppercase-initial, so no user type can collide with them.
-            Ty::Applied(TypeConstructorId::List, la)
-                if la.len() == 1 && !crate::lower::is_heap_ty(&la[0]) =>
-            {
-                "list_scalar".to_string()
-            }
-            Ty::Applied(TypeConstructorId::List, la)
-                if la.len() == 1 && matches!(la[0], Ty::String) =>
-            {
-                "list_str".to_string()
-            }
-            _ if !crate::lower::is_heap_ty(t) => "scalar".to_string(),
-            _ => unreachable!("slot_class admitted only Named/String/Bytes/list/scalar"),
-        }
-    }
-    impl almide_ir::visit::IrVisitor for Finder<'_> {
-        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-            self.check(&e.ty);
-            walk_expr(self, e);
-        }
-    }
-    let mut finder = Finder { pairs: std::collections::BTreeSet::new(), slot_class: &slot_class };
-    let fns = program
-        .functions
-        .iter()
-        .chain(program.modules.iter().flat_map(|m| m.functions.iter()));
-    for f in fns {
-        finder.check(&f.ret_ty);
-        almide_ir::visit::IrVisitor::visit_expr(&mut finder, &f.body);
-    }
-
-    let mut out = String::new();
-    let mut need_vp_list_str = false;
-    for (a, b) in &finder.pairs {
-        need_vp_list_str = need_vp_list_str || a == "list_str" || b == "list_str";
-    }
-    if need_vp_list_str {
-        // The vp-PRIVATE `List[String]` element sweep. The shared
-        // `__drop_list_str` is gated on record/variant FIELD usage
-        // (`program_uses_list_str_drop_field`) and would dangle for a program
-        // whose only `List[String]` owner is a vp slot — a private namespaced
-        // copy needs no cross-generator gate and can never double-define.
-        out.push_str(
-            "fn __drop_vp_list_str(xs: List[String]) -> Unit = {\n  \
-               let h = prim.handle(xs)\n  \
-               if prim.load32(h + 0) == 1 then __drop_vp_list_str_loop(h, prim.load32(h + 4), 0) else ()\n  \
-               prim.rc_dec(h)\n}\n\
-             fn __drop_vp_list_str_loop(h: Int, n: Int, i: Int) -> Unit =\n  \
-               if i >= n then ()\n  \
-               else { prim.rc_dec(prim.load64(h + 12 + i * 8))\n         \
-                      __drop_vp_list_str_loop(h, n, i + 1) }\n",
-        );
-    }
-    for (a, b) in &finder.pairs {
-        let fa = drop_fn_ident(a);
-        let fb = drop_fn_ident(b);
-        let slot_free = |name: &str, off: u32, ident: &str| -> String {
-            if name == "scalar" {
-                // A raw scalar slot owns nothing — freeing it would rc_dec a
-                // non-handle. Emit no free for the slot.
-                return String::new();
-            }
-            if name == "list_str" {
-                return format!(
-                    "    let s{off}: List[String] = prim.load_handle(h + {off})\n    __drop_vp_list_str(s{off})\n"
-                );
-            }
-            if rich.contains(name) || rich_recs.contains(name) {
-                format!(
-                    "    let s{off}: {name} = prim.load_handle(h + {off})\n    __drop_{ident}(s{off})\n"
-                )
-            } else {
-                format!("    prim.rc_dec(prim.load32(h + {off}))\n")
-            }
-        };
-        out.push_str(&format!(
-            "fn __drop_vp_{fa}_{fb}(p: List[Int]) -> Unit = {{\n  \
-               let h = prim.handle(p)\n  \
-               if prim.load32(h + 0) == 1 then {{\n{}{}  }} else ()\n  \
-               prim.rc_dec(h)\n}}\n",
-            slot_free(a, 12, &fa),
-            slot_free(b, 20, &fb),
-        ));
-    }
-    out
-}
-
-/// The RICH-capture env tag for a type name — FNV-1a 64 folded positive. Both
-/// sides of the closure-env RICH class (#1547 shapes 2/3) derive tags through
-/// THIS one function: the lowering stamps `rich_env_tag(elem_name)` into the
-/// capture's wrapper block, and `generate_closure_env_rich_sources` emits the
-/// matching `if tag == <same>` dispatcher arm — no shared registry, no
-/// ordering to keep in sync. A collision between two of one program's type
-/// names would mis-dispatch a free, so the generator PANICS on one (64-bit
-/// FNV over a program's handful of short names — unreachable in practice,
-/// loud if ever).
-pub fn rich_env_tag(name: &str) -> i64 {
-    let mut h: u64 = 0xcbf29ce484222325;
-    for b in name.as_bytes() {
-        h ^= *b as u64;
-        h = h.wrapping_mul(0x100000001b3);
-    }
-    (h & 0x7fff_ffff_ffff_ffff) as i64
-}
-
-/// The RICH-capture env dispatcher `__drop_env_rich` (#1547 shapes 2/3): frees
-/// ONE rich closure capture — a `[tag@12][list-handle@20]` wrapper block whose
-/// list element type needs a per-element recursive free. Tag-dispatches (via
-/// [`rich_env_tag`]) to `$__drop_list_<V>` for a RICH variant element (emitted
-/// unconditionally for every rich variant by `generate_variant_drop_sources`)
-/// or to a `$__drop_caplist_<R>` loop emitted HERE for a recursive-drop record
-/// element (`$__drop_<R>` is likewise unconditional; the record generator's
-/// own `__drop_list_<R>` is field-usage-gated, so the capture side carries its
-/// own loop under a non-colliding name). Arms cover EVERY admissible type —
-/// admission ⊆ generation by construction, never a silent leak-by-missing-arm.
-/// Injected under the same `usage.closures` gate as [`CLOSURE_DROP_SRC`]
-/// (whose rich walk arm calls this); a program with no rich types gets the
-/// tiny always-sound stub. Trusted prim-only routines, leak-loop class.
-pub fn generate_closure_env_rich_sources(
-    type_decls: &[almide_ir::IrTypeDecl],
-) -> String {
-    let variant_names = variant_type_names(type_decls);
-    let all_record_names: std::collections::HashSet<String> = type_decls
-        .iter()
-        .filter(|d| matches!(&d.kind, almide_ir::IrTypeDeclKind::Record { .. }))
-        .map(|d| d.name.as_str().to_string())
-        .collect();
-    let rich_variants: std::collections::BTreeSet<String> = type_decls
-        .iter()
-        .filter(|d| variant_needs_recursive_drop(d, &variant_names, &all_record_names))
-        .map(|d| d.name.as_str().to_string())
-        .collect();
-    let rec_records: std::collections::BTreeSet<String> =
-        recursive_record_drop_names(type_decls).into_iter().collect();
-    // The ROUTED-CELL arm sets (`cell:map_<V>` / `cell:map_rec_<R>` — the #1143
-    // captured `var stats: Map[String, Acc]`): every variant + every ALL-SCALAR
-    // record, the EXACT sets `variant_map_drop_sources` emits `__drop_map_<V>` /
-    // `__drop_map_rec_<R>` for — mirroring `map_named_value_drop`'s admission.
-    let cell_variants: std::collections::BTreeSet<String> = type_decls
-        .iter()
-        .filter(|d| matches!(&d.kind, almide_ir::IrTypeDeclKind::Variant { .. }))
-        .map(|d| d.name.as_str().to_string())
-        .collect();
-    let cell_scalar_recs: std::collections::BTreeSet<String> = type_decls
-        .iter()
-        .filter_map(|d| match &d.kind {
-            almide_ir::IrTypeDeclKind::Record { fields }
-                if fields.iter().all(|f| !crate::lower::is_heap_ty(&f.ty)) =>
-            {
-                Some(d.name.as_str().to_string())
-            }
-            _ => None,
-        })
-        .collect();
-
-    // (name, is_variant), variants first then records, BTree order within each —
-    // deterministic emission; tags themselves are order-free (name hashes).
-    let entries: Vec<(&String, bool)> = rich_variants
-        .iter()
-        .map(|n| (n, true))
-        .chain(rec_records.iter().filter(|n| !rich_variants.contains(*n)).map(|n| (n, false)))
-        .collect();
-    // (map drop suffix, value type name) per cell arm.
-    let cell_entries: Vec<(String, &String)> = cell_variants
-        .iter()
-        .map(|n| (format!("map_{}", drop_fn_ident(n)), n))
-        .chain(cell_scalar_recs.iter().map(|n| (format!("map_rec_{}", drop_fn_ident(n)), n)))
-        .collect();
-    {
-        let mut seen: std::collections::HashMap<i64, String> = std::collections::HashMap::new();
-        let all_tags = entries
-            .iter()
-            .map(|(n, _)| n.to_string())
-            .chain(cell_entries.iter().map(|(sfx, _)| format!("cell:{sfx}")));
-        for n in all_tags {
-            if let Some(prev) = seen.insert(rich_env_tag(&n), n.clone()) {
-                panic!(
-                    "rich_env_tag collision between {prev:?} and {n:?} — \
-                     the closure-env RICH dispatcher cannot distinguish them"
-                );
-            }
-        }
-    }
-
-    let mut out = String::new();
-    if entries.is_empty() && cell_entries.is_empty() {
-        out.push_str("fn __drop_env_rich(wh: Int) -> Unit = prim.rc_dec(wh)\n");
-        return out;
-    }
-    out.push_str(
-        "fn __drop_env_rich(wh: Int) -> Unit = {\n  \
-           if prim.load32(wh + 0) == 1 then {\n    \
-             let tag = prim.load64(wh + 12)\n    ",
-    );
-    for (k, (n, is_variant)) in entries.iter().enumerate() {
-        let fnid = drop_fn_ident(n);
-        let callee = if *is_variant {
-            format!("__drop_list_{fnid}")
-        } else {
-            format!("__drop_caplist_{fnid}")
-        };
-        let kw = if k == 0 { "if" } else { "else if" };
-        out.push_str(&format!(
-            "{kw} tag == {} then {{ let l{k}: List[{n}] = prim.load_handle(wh + 20)\n      {callee}(l{k}) }}\n    ",
-            rich_env_tag(n)
-        ));
-    }
-    // The cell arms: wrapper @20 holds the co-owned CELL; at the cell's last
-    // ref recurse into its @12 inner map via the generated sweep, then free
-    // the cell block.
-    for (k, (sfx, n)) in cell_entries.iter().enumerate() {
-        let kw = if entries.is_empty() && k == 0 { "if" } else { "else if" };
-        out.push_str(&format!(
-            "{kw} tag == {} then {{\n      \
-               let ch{k} = prim.load64(wh + 20)\n      \
-               if prim.load32(ch{k} + 0) == 1 then {{\n        \
-                 let m{k}: Map[String, {n}] = prim.load_handle(ch{k} + 12)\n        \
-                 __drop_{sfx}(m{k})\n      }} else ()\n      \
-               prim.rc_dec(ch{k})\n    }}\n    ",
-            rich_env_tag(&format!("cell:{sfx}"))
-        ));
-    }
-    out.push_str("else ()\n  } else ()\n  prim.rc_dec(wh)\n}\n");
-    for n in rec_records.iter().filter(|n| !rich_variants.contains(*n)) {
-        let fr = drop_fn_ident(n);
-        out.push_str(&format!(
-            "fn __drop_caplist_{fr}(xs: List[{n}]) -> Unit = {{\n  \
-               let h = prim.handle(xs)\n  \
-               if prim.load32(h + 0) == 1 then __drop_caplist_{fr}_loop(h, prim.load32(h + 4), 0) else ()\n  \
-               prim.rc_dec(h)\n}}\n\
-             fn __drop_caplist_{fr}_loop(h: Int, n: Int, i: Int) -> Unit =\n  \
-               if i >= n then ()\n  \
-               else {{ let e: {n} = prim.load_handle(h + 12 + i * 8)\n         __drop_{fr}(e)\n         __drop_caplist_{fr}_loop(h, n, i + 1) }}\n"
-        ));
-    }
-    out
-}
