+++ crates/almide-mir/src/lower/mod_c.rs
@@ -969,6 +969,7 @@ include!("drop_sources.rs");
 include!("variant_drop_field_frees.rs");
 include!("drop_sources_b.rs");
 include!("drop_sources_c.rs");
+include!("drop_sources_d.rs");
 include!("repr_sources.rs");
 include!("repr_sources_b.rs");
 include!("repr_sources_c.rs");
